py/pyproject.toml
[tool.poetry]
name = "davidkhala.gcp.core"
version = "0.0.0"
description = ""
authors = ["David Liu <david-khala@hotmail.com>"]
readme = "README.md"
packages = [{include = "davidkhala"}]

[tool.poetry.dependencies]
python = "^3.12"
google-auth = "*"

[tool.poetry.group.dev.dependencies]
pytest = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
